Move unused prepare_restraints to attic
