+++ tests/common.py
@@ -16,7 +16,7 @@ from rdkit import Chem
 from timemachine.constants import ONE_4PI_EPS0
 from timemachine.ff import Forcefield
 from timemachine.lib import potentials
-from timemachine.potentials import bonded, generic, nonbonded
+from timemachine.potentials import generic, nonbonded
 
 
 @contextlib.contextmanager
@@ -148,29 +148,6 @@ def prepare_nb_system(x, E, lambda_plane_idxs, lambda_offset_idxs, p_scale, cuto
     )
 
     return params, ref_total_energy, test_potential
-
-
-def prepare_restraints(x, B, precision):
-
-    assert x.ndim == 2
-    N = x.shape[0]
-    # D = x.shape[1]
-
-    atom_idxs = np.arange(N)
-
-    params = np.random.randn(B, 3).astype(np.float64)
-
-    bond_idxs = []
-    for _ in range(B):
-        bond_idxs.append(np.random.choice(atom_idxs, size=2, replace=False))
-    bond_idxs = np.array(bond_idxs, dtype=np.int32)
-
-    lambda_flags = np.random.randint(0, 2, size=(B,)).astype(np.int32)
-
-    custom_restraint = potentials.Restraint(bond_idxs, params, lambda_flags, precision=precision)
-    restraint_fn = functools.partial(bonded.restraint, box=None, lamb_flags=lambda_flags, bond_idxs=bond_idxs)
-
-    return (params, restraint_fn), custom_restraint
 
 
 def hilbert_sort(conf, D):
